+++ options/db_options.h
@@ -85,6 +85,7 @@ struct ImmutableDBOptions {
   bool two_write_queues;
   bool manual_wal_flush;
   bool atomic_flush;
+  bool avoid_unnecessary_blocking_io;
 };
 
 struct MutableDBOptions {