+++ options/options.cc
@@ -562,7 +562,6 @@ ReadOptions::ReadOptions()
       managed(false),
       total_order_seek(false),
       prefix_same_as_start(false),
-      background_purge_on_iterator_cleanup(false),
       ignore_range_deletions(false),
       aio_concurrency(32),
       iter_start_seqnum(0) {}
@@ -580,7 +579,6 @@ ReadOptions::ReadOptions(bool cksum, bool cache)
       managed(false),
       total_order_seek(false),
       prefix_same_as_start(false),
-      background_purge_on_iterator_cleanup(false),
       ignore_range_deletions(false),
       aio_concurrency(32),
       iter_start_seqnum(0) {}